cleaning up exports for config
