+++ packages/analytics/src/config.ts
@@ -1,0 +1,15 @@
+import { NextConfig } from "next";
+
+type AnalyticsConfig = {
+  debug?: boolean;
+};
+
+const withAnalytics = (
+  pluginConfig: AnalyticsConfig
+): ((nextConfig: NextConfig) => NextConfig) => {
+  return function withAnalytics(nextConfig: NextConfig) {
+    return nextConfig;
+  };
+};
+
+export default withAnalytics;
