+++ packages/cms/src/index.ts
@@ -1,5 +1,2 @@
-export { startupkitCMS } from "./plugin";
 export { CMSProvider } from "./provider";
-export * from "./utils";
-
-export const questions = [];
+export * from "./lib/utils";
