+++ template/next.config.js
@@ -1,5 +1,8 @@
+import withAnalytics from "@startupkit/analytics/config";
+import withCMS from "@startupkit/cms/config";
+
 /** @type {import('next').NextConfig} */
-const nextConfig = {
+const baseConfig = {
   env: {
     STRIPE_PUBLISHABLE_KEY: process.env.STRIPE_PUBLISHABLE_KEY,
     WORKOS_API_KEY: process.env.WORKOS_API_KEY,
@@ -8,4 +11,8 @@ const nextConfig = {
   },
 };
 
-module.exports = nextConfig;
+const combine = (...plugins) => {
+  return plugins.reduce((acc, plugin) => plugin(acc), config);
+};
+
+module.exports = combine(withAnalytics(), withCMS(), baseConfig);
